+++ backend/src/notifications/notifications.module.ts
@@ -1,12 +1,20 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { Flower } from '../flowers/flower.entity';
 import { PushModule } from '../push/push.module';
+import { StorageModule } from '../storage/storage.module';
+import { UsersModule } from '../users/users.module';
 import { Notification } from './notification.entity';
 import { NotificationsController } from './notifications.controller';
 import { NotificationsService } from './notifications.service';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Notification]), PushModule],
+  imports: [
+    TypeOrmModule.forFeature([Notification, Flower]),
+    PushModule,
+    UsersModule,
+    StorageModule,
+  ],
   controllers: [NotificationsController],
   providers: [NotificationsService],
   exports: [NotificationsService],